requirements.txt
pathway
kafka-python
fastapi
uvicorn
requests
textblob
numpy
openai
pdf2image
pillow
unstructured